// src/schema/user/tests/updateMyUserData/updateMyUserData.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type UpdateMyUserDataMutationVariables = Types.Exact<{
  input: Types.UpdateUserDataInput;
}>;


export type UpdateMyUserDataMutation = { __typename?: 'Mutation', updateMyUserData: { __typename?: 'User', id: string, userData: { __typename?: 'UserData', city: string, countryOfResidence: string, organizationName: string | null, roleInOrganization: string | null, worksInOrganization: boolean } | null } };


export const UpdateMyUserData = gql`
    mutation UpdateMyUserData($input: updateUserDataInput!) {
  updateMyUserData(input: $input) {
    id
    userData {
      city
      countryOfResidence
      organizationName
      roleInOrganization
      worksInOrganization
    }
  }
}
    `;